Only add Series if series exists

+++ inc/template-tags.php
@@ -141,13 +141,15 @@ function twentysixteen_entry_taxonomies() {
 //			_x( 'Categories', 'Used before category names.', 'twentysixteen' ),
 //			$categories_list
 //		);
-//	}
-	$series_list = get_the_term_list( get_the_ID(), 'series', _x( ', ', 'Used between list items, there is a space after the comma.', 'twentysixteen' ) );
-	if ( $series_list ) {
-		printf( '<span class="series-links"><span class="screen-reader-text">%1$s </span>%2$s</span>',
-			_x( 'Series', 'Used before series names.', 'twentysixteen' ),
-			$series_list
-		);
+	//	}
+	if ( taxonomy_exists( 'series' ) ) {
+		$series_list = get_the_term_list( get_the_ID(), 'series', _x( ', ', 'Used between list items, there is a space after the comma.', 'twentysixteen' ) );
+		if ( $series_list ) {
+			printf( '<span class="series-links"><span class="screen-reader-text">%1$s </span>%2$s</span>',
+				_x( 'Series', 'Used before series names.', 'twentysixteen' ),
+				$series_list
+			);
+		}
 	}
 	$tags_list = get_the_tag_list( '', _x( ', ', 'Used between list items, there is a space after the comma.', 'twentysixteen' ) );
 	if ( $tags_list ) {
